cwk1 _ mysite/requirements.txt
django
requests
djangorestframework